Revert "chore: fix clippy"
This reverts commit 9b67ee682c7bcb4d1bd845cce92c23456e905e69.

+++ src/gridfs/download.rs
@@ -192,7 +192,7 @@ impl GridFsBucket {
         if n != file.n() {
             return Err(ErrorKind::GridFs(GridFsErrorKind::WrongNumberOfChunks {
                 actual_number: n,
-                expected_number: file.n(),
+                expected_number: file.n() as u32,
             })
             .into());
         }
